refactor: minor fixes
