+++ src/platform/api/src/infrastructure/repositories/sqlite/SQLiteGoogleAuthRepository.ts
@@ -1,5 +1,5 @@
 import { GoogleAuth } from "../../../domain/entities/GoogleAuth";
-import { IGoogleAuthRepository } from "../../../domain/repositories/IGoogleAuthRepositroy";
+import { IGoogleAuthRepository } from "../../../domain/repositories/IGoogleAuthRepository";
 import { SQLiteBaseRepository } from "./SQLiteBaseRepository";
 
 export class SQLiteGoogleAuthRepository extends SQLiteBaseRepository<GoogleAuth> implements IGoogleAuthRepository {
